# fifocache: cherry-pick ghost queue changes and optimize