--- a/pkg/fileservice/disk_cache.go
+++ b/pkg/fileservice/disk_cache.go
@@ -119,6 +119,7 @@ func NewDiskCache(
 					)
 				}
 			},
+			-1,
 		),
 	}
 	ret.updatingPaths.Cond = sync.NewCond(new(sync.Mutex))