## What type of PR is this?

- [ ] API-change
- [ ] BUG
- [x] Improvement
- [ ] Documentation
- [ ] Feature
- [ ] Test and CI
- [ ] Code Refactoring

## Which issue(s) this PR fixes:

issue issue https://github.com/matrixorigin/matrixone/issues/21654 https://github.com/matrixorigin/matrixone/issues/21589

## What this PR does / why we need it:
Cherry-pick ghost queue changes by Eric.
And optimize the ghost queue with Queue.

benchmark results:
```
goos: linux
goarch: amd64
pkg: github.com/matrixorigin/matrixone/pkg/fileservice/fifocache
cpu: AMD Ryzen 9 7900 12-Core Processor             
                      │ /home/reus/mo/pkg/fileservice/fifocache/b1 │                 b2                  │
                      │                   sec/op                   │    sec/op     vs base               │
SequentialSet-24                                      139.5n ± 11%   232.1n ±  1%  +66.28% (p=0.002 n=6)
ParallelSet-24                                        18.12n ±  2%   23.73n ± 24%  +30.96% (p=0.037 n=6)
Get-24                                                34.67n ±  3%   34.62n ±  1%        ~ (p=0.589 n=6)
ParallelGet-24                                        7.571n ±  6%   7.670n ±  5%        ~ (p=0.589 n=6)
ParallelGetOrSet-24                                   13.76n ±  2%   13.73n ±  7%        ~ (p=0.667 n=6)
ParallelEvict-24                                      160.6n ±  7%   147.0n ±  6%   -8.47% (p=0.002 n=6)
ShardCacheKey-24                                      21.32n ±  3%   21.08n ±  1%        ~ (p=0.180 n=6)
EnsureNBytesAndSet-24                                 16.44n ± 11%   16.43n ±  1%        ~ (p=0.589 n=6)
DataCacheSet-24                                       255.1n ±  4%   377.4n ±  4%  +47.94% (p=0.002 n=6)
DataCacheGet-24                                       36.73n ±  1%   36.57n ±  7%        ~ (p=0.177 n=6)
geomean                                               37.01n         41.20n        +11.34%

                      │ /home/reus/mo/pkg/fileservice/fifocache/b1 │                  b2                  │
                      │                    B/op                    │     B/op      vs base                │
SequentialSet-24                                    32.00 ±   0%     32.00 ±   0%       ~ (p=1.000 n=6) ¹
ParallelSet-24                                      1.000 ±   0%     1.000 ±   0%       ~ (p=1.000 n=6) ¹
Get-24                                              0.000 ±   0%     0.000 ±   0%       ~ (p=1.000 n=6) ¹
ParallelGet-24                                      0.000 ±   0%     0.000 ±   0%       ~ (p=1.000 n=6) ¹
ParallelGetOrSet-24                                 0.000 ±   0%     0.000 ±   0%       ~ (p=1.000 n=6) ¹
ParallelEvict-24                                    1.000 ± 400%     1.000 ± 100%       ~ (p=0.455 n=6)
ShardCacheKey-24                                    0.000 ±   0%     0.000 ±   0%       ~ (p=1.000 n=6) ¹
EnsureNBytesAndSet-24                               0.000 ±   0%     0.000 ±   0%       ~ (p=1.000 n=6) ¹
DataCacheSet-24                                     118.0 ±   0%     119.0 ±   0%  +0.85% (p=0.002 n=6)
DataCacheGet-24                                     0.000 ±   0%     0.000 ±   0%       ~ (p=1.000 n=6) ¹
geomean                                                          ²                 +0.08%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean

                      │ /home/reus/mo/pkg/fileservice/fifocache/b1 │                 b2                  │
                      │                 allocs/op                  │ allocs/op   vs base                 │
SequentialSet-24                                      1.000 ± 0%     1.000 ± 0%        ~ (p=1.000 n=6) ¹
ParallelSet-24                                        0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=6) ¹
Get-24                                                0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=6) ¹
ParallelGet-24                                        0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=6) ¹
ParallelGetOrSet-24                                   0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=6) ¹
ParallelEvict-24                                      0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=6) ¹
ShardCacheKey-24                                      0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=6) ¹
EnsureNBytesAndSet-24                                 0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=6) ¹
DataCacheSet-24                                       4.000 ± 0%     5.000 ± 0%  +25.00% (p=0.002 n=6)
DataCacheGet-24                                       0.000 ± 0%     0.000 ± 0%        ~ (p=1.000 n=6) ¹
geomean                                                          ²                +2.26%               ²
¹ all samples are equal
² summaries must be >0 to compute geomean

```