@reusee 2025-04-30
> How about you also make a configuration param, so that we can protect, enable, disable this change?
> 
> Thank you

增加了，可以通过配置文件调整 ghost queue 大小。为 0 时禁用。